Started with support for glyphnames
If a font contains Adobe glyphnames, record them as well,
because they might contain important information needed for
text extraction, especially if the Unicode Private Use Area
is being used.

Suppport is not yet complete. The resolving of glyphnames to
Unicode codepoints will have to be refined in order to support
names like "a.sc".